LayoutTests/fast/line-grid/line-align-parsing-expected.txt

Test parsing of the CSS line-snap property.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS test("-webkit-line-align: none") is "none"
PASS test("-webkit-line-align: edges") is "edges"
PASS test("-webkit-line-align: ;") is ""
PASS test("-webkit-line-align: 1") is ""
PASS test("-webkit-line-align: 1.2") is ""
PASS test("-webkit-line-align: -1") is ""
PASS test("-webkit-line-align: 12px") is ""
PASS testComputedStyle("none") is "none"
PASS testComputedStyle("") is "none"
PASS testComputedStyle("12px") is "none"
PASS testInherited("none") is "none"
PASS testInherited("edges") is "edges"
PASS successfullyParsed is true

TEST COMPLETE
